[FEAT] Fruit Wearable quest
