+++ src/features/game/expansion/components/Water.tsx
@@ -14,7 +14,7 @@ import { MapPlacement } from "./MapPlacement";
 import { Snorkler } from "./water/Snorkler";
 import { SharkBumpkin } from "./water/SharkBumpkin";
 import { Arcade } from "features/community/arcade/Arcade";
-import { FarmerQuest } from "features/island/farmerQuest/FarmerQuest";
+import { FruitQuest } from "features/island/farmerQuest/FruitQuest";
 
 import { merchantAudio } from "lib/utils/sfx";
 import { ProjectDignityFrogs } from "features/community/components/ProjectDignityFrogs";
@@ -122,7 +122,7 @@ export const Water: React.FC<Props> = ({ level }) => {
         </div>
       </MapPlacement>
 
-      <FarmerQuest />
+      <FruitQuest />
 
       <MapPlacement x={-20} y={-15} width={6}>
         <img
