+++ src/features/game/types/quests.ts
@@ -1,12 +1,11 @@
+import { GameState } from "../types/game";
 import { BumpkinItem } from "./bumpkin";
-import { GameState } from "./game";
 
 export type QuestName =
   | "Farmer Quest 1"
   | "Fruit Quest 1"
   | "Fruit Quest 2"
   | "Fruit Quest 3"
-  | "Fruit Quest 4"
   | "Lunar New Year Quest 1"
   | "Lunar New Year Quest 2";
 
@@ -31,41 +30,29 @@ export const QUESTS: Record<QuestName, Quest> = {
     progress: (gameState: GameState) =>
       gameState.bumpkin?.activity?.["Blueberry Harvested"] || 0,
     requirement: 10,
-    wearable: "Red Farmer Shirt",
-    deadline: new Date(Date.now() + 10000000000).toISOString(),
+    wearable: "Fruit Picker Shirt",
   },
   "Fruit Quest 2": {
-    description: "Harvest 100 Oranges",
+    description: "Harvest 100 Orange",
     progress: (gameState: GameState) =>
       gameState.bumpkin?.activity?.["Orange Harvested"] || 0,
     requirement: 100,
-    wearable: "Farmer Overalls",
-    deadline: new Date(Date.now() + 10000000000).toISOString(),
+    wearable: "Fruit Picker Apron",
   },
   "Fruit Quest 3": {
-    description: "Harvest 200 Apples",
+    description: "Harvest 750 Apples",
     progress: (gameState: GameState) =>
       gameState.bumpkin?.activity?.["Apple Harvested"] || 0,
-    requirement: 200,
-    wearable: "Farmer Hat",
-    deadline: new Date(Date.now() + 10000000000).toISOString(),
+    requirement: 750,
+    wearable: "Fruit Bowl",
   },
-  "Fruit Quest 4": {
-    description: "Harvest 1200 Blueberries",
-    progress: (gameState: GameState) =>
-      gameState.bumpkin?.activity?.["Blueberry Harvested"] || 0,
-    requirement: 1200,
-    wearable: "Farmer Pitchfork",
-    deadline: new Date(Date.now() + 10000000000).toISOString(),
-  },
-
   "Lunar New Year Quest 1": {
     description: "Collect 25 Red Envelopes",
     progress: (gameState: GameState) =>
       gameState.inventory["Red Envelope"]?.toNumber() || 0,
     requirement: 25,
     wearable: "China Town Background",
-    deadline: new Date("2023-01-29T00:00:00").toISOString(),
+    deadline: new Date(Date.now() + 10000000000).toISOString(),
   },
 
   "Lunar New Year Quest 2": {
@@ -74,7 +61,7 @@ export const QUESTS: Record<QuestName, Quest> = {
       gameState.inventory["Red Envelope"]?.toNumber() || 0,
     requirement: 125,
     wearable: "Lion Dance Mask",
-    deadline: new Date("2023-01-29T00:00:00").toISOString(),
+    deadline: new Date(Date.now() + 10000000000).toISOString(),
   },
 };
 
@@ -92,7 +79,6 @@ export const BUMPKIN_QUEST_IDS: Record<QuestName | CompletedQuestName, number> =
     "Fruit Quest 1": 100005,
     "Fruit Quest 2": 100006,
     "Fruit Quest 3": 100007,
-    "Fruit Quest 4": 100008,
     "Lunar New Year Quest 1": 100009,
     "Lunar New Year Quest 2": 100010,
   };
